+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [1.0.9] - 2018-10-22
+[feature] added filter by topic in subscribers
+[fix] mqtt 5 settings init values
+[fix] selecting info in message card
 ### [1.0.8] - 2018-10-22
 [fix] stop client with active subscribes
 ### [1.0.7] - 2018-10-18